refactor(app): move `context` to `/profile`
