+++ app/profile/providers.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { PlaybackStateProvider } from '@app/context/playback-state'
+import { PlaybackStateProvider } from '@app/profile/context/playback-state'
 import { LayoutProps } from '@app/types'
 
 export function ProfileProviders({ children }: LayoutProps) {
